Bug 1181913 - fix data race in AudioSink::mWritten. r=kinetik
UltraBlame original commit: 2e5a81baa18b5f1aaeab4e7e72d5ac63a77f58ca

+++ dom/media/AudioSink.cpp
@@ -199,7 +199,8 @@ AudioSink::AudioLoop()
     CheckedInt64 sampleTime = UsecsToFrames(AudioQueue().PeekFront()->mTime, mInfo.mRate);
 
     
-    CheckedInt64 playedFrames = UsecsToFrames(mStartTime, mInfo.mRate) + mWritten;
+    CheckedInt64 playedFrames = UsecsToFrames(mStartTime, mInfo.mRate) +
+                                static_cast<int64_t>(mWritten);
 
     CheckedInt64 missingFrames = sampleTime - playedFrames;
     if (!missingFrames.isValid() || !sampleTime.isValid()) {
